Move RunGuardian to distinct package
