+++ src/main/scala/edu/ie3/osmogrid/guardian/run/SubGridHandling.scala
@@ -4,13 +4,14 @@
  * Research group Distribution grid planning and operation
  */
 
-package edu.ie3.osmogrid.guardian
+package edu.ie3.osmogrid.guardian.run
 
 import akka.actor.typed.ActorRef
 import edu.ie3.datamodel.models.input.container.SubGridContainer
 import edu.ie3.datamodel.utils.ContainerUtils
 import edu.ie3.osmogrid.cfg.OsmoGridConfig
-import edu.ie3.osmogrid.guardian.SubGridHandling.assignSubnetNumbers
+import edu.ie3.osmogrid.guardian.run.RunGuardian
+import edu.ie3.osmogrid.guardian.run.SubGridHandling.assignSubnetNumbers
 import edu.ie3.osmogrid.io.output.ResultListener
 import org.slf4j.Logger
